master: Full overhall. Everything is working except symlinks.

+++ lib/helpers.py
@@ -26,14 +26,16 @@ if '--dry_run' in sys.argv:
 # Any json we use should (for now) should not use spaces or - in the key
 def dict_to_obj(blueprint_dict):
   blueprint_list = list(blueprint_dict.keys())
-  proto = namedtuple('something', blueprint_list)
+  proto = namedtuple('dynamic_holder', blueprint_list)
   result_dict = proto(**blueprint_dict)
   return result_dict
 
-#TODO(xnz): We need these paths, should be abstracted to a json file.
+#TODO(xnz): Put BASE_DIR in other file , file_paths dict comp from it. if no file then use home
+BASE_DIR = os.environ['HOME'] + '/dev/sandbox'
 file_paths = {
-  'config_file': 'path',
-  'addons_path': 'path'
+  'base_dir': BASE_DIR,  # + is temps
+  'addons_config': BASE_DIR + '/clither_custom/config.json',
+  'rcs_path': BASE_DIR + '/clither_custom/rcs',
 }
 
 paths = dict_to_obj(file_paths)
@@ -46,11 +48,25 @@ def json_to_obj(json_file_path):
   return data
 
 def create_directory(dir):
-  if dry_run('create dir ' + dir):
+  if os.path.exists(dir):
     return
 
-  if not os.path.exists(dir):
-    os.makedirs(dir)
+  msg = 'mkdir ' + dir
+  if dry_run(msg):
+    return
+
+  os.makedirs(dir)
+  print(msg)
+
+def mk_clither_custom_dirs():
+  dirs = (
+    '',  # base dir
+    'rcs',
+    'addons',
+  )
+  for dir in dirs:
+    create_directory('{0}/clither_custom/{1}'.format(paths.base_dir, dir))
+
 
 def backup_file(file_path):
   """Backup file at a given path.
@@ -58,6 +74,10 @@ def backup_file(file_path):
   args:
     file_path: (str) Path to backup file.
   """
+  # msg = 'backup_file: ' + file_path
+  # if dry_run(msg):
+  #   return
+
   if os.path.isfile(file_path):
     # get file name and containing directory
     dir_path = os.path.dirname(file_path)
@@ -71,8 +91,32 @@ def backup_file(file_path):
     #   e.g. zshrc_1234567890
     timestamp = int(time.time())
     dst_file_path = '{0}/{1}_{2}'.format(backup_path, file_name, timestamp)
-    print('{0} backed up to: {1}'.format(file_name, dst_file_path))
+    # print('{0} backed up to: {1}'.format(file_name, dst_file_path))
+
+    msg = 'backup_file: cp {0} {1}'.format(file_path, dst_file_path)
+    if dry_run(msg):
+      return
     os.rename(file_path, dst_file_path)
+    print(msg)
+
+def clear_file(file_path):
+  msg = 'clear_file: ' + file_path
+  if dry_run(msg):
+    return
+
+  #TODO(xnz): Check for better way to do this.
+  with open(file_path, 'w') as file:
+    file.write('')
+  print(msg)
+
+def append_to_file(file_path, string):
+  msg = 'append_to_file: echo {0} >> {1}'.format(string, file_path)
+  if dry_run(msg):
+    return
+
+  with open(file_path, 'a') as file:
+    file.write(string + '\n')
+  print(msg)
 
 def dry_run(msg):
   """Return True and print msg if dry_run flag is set, else return False."""
@@ -87,7 +131,18 @@ def run_cmd(cmd):
   args:
     cmd: (str) The command to run.
   """
-  if dry_run('run cmd: ' + cmd):
+  msg = 'run cmd: ' + cmd
+  if dry_run(msg):
     return
   #TODO(xnz): mv this to a subprocess
   os.system(cmd)
+  print(msg)
+
+def create_symlink(src, dst):
+  # check if already exists
+  msg = 'create symlink: ln -s {0} {1}'.format(src, dst)
+  if dry_run(msg):
+    return
+
+  os.symlink(src, dst)
+  print(msg)
